feat(VsCheckboxSet): apply code review
