+++ packages/vlossom/src/composables/index.ts
@@ -3,4 +3,5 @@ export * from './custom-style-composable';
 export * from './form-provide-composable';
 export * from './input-composable';
 export * from './input-form-composable';
+export * from './input-option-composable';
 export * from './responsive-composable';
